refactor(core): move Region's eviction into region/evict.rs
A pure move of the eight methods that decide what leaves a region:
`make_room`, `would_roll_off`, `release_oldest`, `evict_lru_entry`,
`enforce_sliding_window`, `turn_group_size_at` and `remove_oldest`, into a
child module with its own `impl Region`. Nothing else changes; the tests
stay in `mod.rs`. Verified as a line-multiset move against main.
